Drug listing

+++ app/src/main/java/com/dariomartin/easygrow/data/dto/DrugDTO.kt
@@ -4,8 +4,8 @@ data class DrugDTO(
     var id: String = "",
     val name: String = "",
     val pharmacy: String = "",
-    val amountOfDrugMg: Int = 0,
-    val volumeOfDrugMl: Int = 0,
-    val cartridgeVolumeMl: Int = 0,
+    val amountOfDrugMg: Float = 0F,
+    val volumeOfDrugMl: Float = 0F,
+    val cartridgeVolumeMl: Float = 0F,
     val url: String = ""
 )
